+++ Raspberry/rechenblock.h
@@ -7,7 +7,7 @@ class Rechenblock
 public:
     Rechenblock();
     std::list<Rechenblock*> eingaenge;
-    virtual float get_data(uint32_t time)=0;
+    virtual double get_data(uint32_t time)=0;
     virtual uint32_t newest()=0;
 };
