+++ Raspberry/rechenbloecke.cpp
@@ -1,8 +1,8 @@
 #include "rechenbloecke.h"
 
-float Time::get_data(uint32_t time)
+double Time::get_data(uint32_t time)
 {
-    return (float)time / Control::samplefreq;
+    return (double)time / Control::samplefreq;
 }
 uint32_t Time::newest()
 {
